Module rQuery and rRecord intorduction
rQuery is the script that queries reddit for specific post data. rRecord
stores that data in a MySQL database. To run either module you need to have
the modules csv and os installed. Additionally, rQuery is dependent on praw,
and rRecord is dependent on mysqldb. You must be able to connect to a mysql
database. You must have a database called rQuery_Results in MySQL.